Captures output for update functions
Modifies update functions to capture the output of external commands (choco, winget, wsl) for logging and debugging purposes.

This change ensures that the output of these commands, which may include error messages or progress information, is stored in the `$result.Output` variable, rather than directly outputting to the console. This helps to keep the console clean and allows for more detailed logging and analysis of the update process.

Also, addresses issues where errors were not being captured correctly in the WSL update function and fixes function not found messages for clarity.